refactor(api): move base to separate file with no eslint validation
