+++ apps/api/src/app/shared/helpers/base62/base62-alphabet.const.ts
@@ -1,0 +1,5 @@
+/**
+ * Base62 alphabet
+ * Modifying this alphabet is prohibited as it would invalidate existing encoded data
+ */
+const BASE62_ALPHABET = '0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz';
