+++ apps/api/src/app/shared/helpers/base62/index.ts
@@ -1,0 +1,1 @@
+export * from './base62';
